refactor(EndpointsList): split state from render
