+++ webapp/src/routes/EndpointsList/index.js
@@ -1,6 +1,5 @@
 /* eslint camelcase: 0 */
-import React, { useEffect, useState } from 'react'
-import { useLazyQuery } from '@apollo/client'
+import React from 'react'
 import { useTranslation } from 'react-i18next'
 import { makeStyles } from '@mui/styles'
 import Card from '@mui/material/Card'
@@ -14,7 +13,7 @@ import Select from '@mui/material/Select'
 import Pagination from '@mui/material/Pagination'
 import moment from 'moment'
 
-import { PRODUCERS_QUERY } from '../../gql'
+import useEndpointsState from '../../hooks/customHooks/useEndpointsState'
 
 import EndpointsTable from './EndpointsTable'
 import styles from './styles'
@@ -26,61 +25,10 @@ const limitOptions = [10, 20, 30, 50, 80, 130]
 const EndpointsList = () => {
   const classes = useStyles()
   const { t } = useTranslation('endpointsListRoute')
-  const [load, { loading, data }] = useLazyQuery(PRODUCERS_QUERY)
-  const [producers, setProducers] = useState([])
-  const [updatedAt, setUpdatedAt] = useState()
-  const [pagination, setPagination] = useState({
-    page: 1,
-    limit: 80,
-    totalPages: 0,
-  })
-
-  useEffect(() => {
-    load({
-      variables: {
-        offset: (pagination.page - 1) * pagination.limit,
-        limit: pagination.limit,
-        where: { endpoints: { _neq: { api: [], ssl: [], p2p: [] } } },
-      },
-    })
-    // eslint-disable-next-line
-  }, [pagination.page, pagination.limit])
-
-  useEffect(() => {
-    if (!data?.info) return
-
-    setPagination((prev) => ({
-      ...prev,
-      totalPages: Math.ceil(data?.info.producers?.count / pagination.limit),
-    }))
-    // eslint-disable-next-line
-  }, [data?.info, pagination.limit])
-
-  useEffect(() => {
-    if (!data?.producers) return
-
-    setProducers(
-      data.producers.map((producer) => ({
-        name:
-          producer.bp_json?.org?.candidate_name ||
-          producer?.bp_json?.org?.organization_name ||
-          producer?.owner,
-        endpoints: producer.endpoints,
-      })),
-    )
-
-    if (!data.producers?.[0]?.updated_at) return
-
-    setUpdatedAt(data.producers[0].updated_at)
-    // eslint-disable-next-line
-  }, [data?.producers])
-
-  const handleOnPageChange = (_, page) => {
-    setPagination((prev) => ({
-      ...prev,
-      page,
-    }))
-  }
+  const [
+    { loading, pagination, producers, updatedAt },
+    { handleOnPageChange, setPagination },
+  ] = useEndpointsState()
 
   return (
     <Card>
